Move to enums directory
