+++ src/Api/Cups/Response.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Rawilk\Printing\Api\Cups;
 
 use Rawilk\Printing\Api\Cups\Enums\AttributeGroupTag;
+use Rawilk\Printing\Api\Cups\Enums\TypeTag;
+use Rawilk\Printing\Api\Cups\Enums\Version;
 use Rawilk\Printing\Api\Cups\Exceptions\UnknownType;
 use Rawilk\Printing\Drivers\Cups\Entity\Printer;
 use Rawilk\Printing\Drivers\Cups\Entity\PrintJob;
