+++ src/Api/Cups/Types/MimeMedia.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Rawilk\Printing\Api\Cups\Types;
 
 use Rawilk\Printing\Api\Cups\Types\Primitive\Text;
-use Rawilk\Printing\Api\Cups\TypeTag;
+use Rawilk\Printing\Api\Cups\Enums\TypeTag;
 
 class MimeMedia extends Text
 {
