+++ src/Api/Cups/Types/Primitive/OctetString.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Rawilk\Printing\Api\Cups\Types\Primitive;
 
-use Rawilk\Printing\Api\Cups\TypeTag;
+use Rawilk\Printing\Api\Cups\Enums\TypeTag;
 
 class OctetString extends Text
 {
