+++ src/Api/Cups/Types/Primitive/Text.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Rawilk\Printing\Api\Cups\Types\Primitive;
 
 use Rawilk\Printing\Api\Cups\Type;
-use Rawilk\Printing\Api\Cups\TypeTag;
+use Rawilk\Printing\Api\Cups\Enums\TypeTag;
 
 class Text extends Type
 {
